# perf: make db calls during applyRound non-blocking